backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
PyPDF2==3.0.1
pikepdf==9.0.0
PyMuPDF==1.27.2.2
pdf2image==1.17.0
Pillow==10.4.0
python-docx==1.1.2
reportlab==4.2.2
aiofiles==24.1.0
cryptography==43.0.0
